ENH: Modularising the Codebase
- Moved some endpoints to new blueprints in modules
- Blueprints avoid the need of passing around aliases to the flask app
- Separated helper methods, resume page, and error page
- Work in progress
