+++ site-flask-mustache/www/src/_base.py
@@ -1,0 +1,38 @@
+TWO_WEEKS_SECONDS = 1_209_600;
+
+THEME_CLASSES = {
+   "default": {
+      "body": "darkonlight",
+      "table": "table-light",
+      "btncls_cookies_store": "btn-outline-primary",
+      "btncls_cookies_erase": "btn-outline-danger",
+      "navbar": "navbar-light bg-light",
+   },
+   "light_on_dark": {
+      "body": "lightondark",
+      "table": "table-dark",
+      "btncls_cookies_store": "btn-outline-primary",
+      "btncls_cookies_erase": "btn-outline-danger",
+      "navbar": "navbar-dark bg-dark",
+   },
+   "dark_on_light": {
+      "body": "darkonlight",
+      "table": "table-light",
+      "btncls_cookies_store": "btn-outline-primary",
+      "btncls_cookies_erase": "btn-outline-danger",
+      "navbar": "navbar-light bg-light",
+   },
+};
+
+BASE_STYLESHEETS = [
+   { "cssname": "/css/bootstrap.min.css", },
+   { "cssname": "/css/index.css", },
+];
+
+BASE_SCRIPTS = [
+   { "scriptname": "/js/jquery.min.js", },
+   { "scriptname": "/js/bootstrap.bundle.min.js", },
+   { "scriptname": "/js/d3.min.js", },
+   { "scriptname": "/js/cookie.js", },
+   { "scriptname": "/js/index.js", },
+];
